Refactor: triggerInputError의 import를 일관화
